# Conflicts:
#	folib-storage/folib-storage-layout-providers/folib-storage-debian-layout-provider/src/main/java/com/veadan/folib/indexer/DebianIncrementalIndexer.java
#	folib-web-core/src/main/java/com/veadan/folib/components/syncartifact/DebianSyncArtifactProvider.java
#	folib-web-core/src/main/java/com/veadan/folib/components/syncartifact/RawSyncArtifactProvider.java
#	folib-web-core/src/main/java/com/veadan/folib/services/impl/ArtifactWebServiceImpl.java
#	folib-web-vue/src/views/Storage/components/Debian/DebianBatchUpload.vue

+++ folib-storage/folib-storage-api/src/main/java/com/veadan/folib/providers/io/StorageFileSystemProvider.java
@@ -174,10 +174,10 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath repositoryPath = (RepositoryPath) path;
-        if (!Files.exists(repositoryPath.getTarget())) {
-            throw new NoSuchFileException(unwrap(repositoryPath).toString());
-        }
-
+//        if (!Files.exists(repositoryPath.getTarget())) {
+//            throw new NoSuchFileException(unwrap(repositoryPath).toString());
+//
+//        }
         if (!Files.isDirectory(repositoryPath)) {
             doDeletePath(repositoryPath, force, true);
 
@@ -282,7 +282,9 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath trashPath = getTrashPath(repositoryPath);
-
+        if (!Files.exists(repositoryPath)) {
+            return;
+        }
         try {
             Files.move(repositoryPath.getTarget(),
                     trashPath.getTarget(),
